split out package validation so that it can be used on server.
The validation code is no longer in ExtensionManagerDomain itself.
