+++ src/extensibility/node/README.md
@@ -1,0 +1,3 @@
+## Brackets Extensibility
+
+This code is used by Brackets to implement its extension management features. It is likely not useful to anyone not working with Brackets extensions.
