+++ src/extensibility/node/package.json
@@ -1,5 +1,7 @@
 {
     "name": "brackets-extensibility",
+    "description": "Used in the management of Brackets extensions",
+    "version": "0.22.0",
     "dependencies": {
         "unzip": "0.1.x",
         "semver": ">1.1.0 <2.0.0",
